Feature/painting widget

+++ src/paintings/paintings.service.ts
@@ -12,9 +12,11 @@ import { Painting } from './models/painting.model'
 import { StorageService } from '../common/services/storage.service'
 import { getFileNameFromUrl } from '../utils'
 import { Artist } from '../artists/models/artist.model'
+import { Attributes } from '../attributes/models/attributes.model'
 import { parsePriceRange } from '../utils/parsePriceRange'
 import { parseSizeList } from '../utils/parseSizeList'
 import { Sequelize } from 'sequelize-typescript'
+import { PaintingAttributes } from './models/painting-attributes.model'
 
 @Injectable()
 export class PaintingsService {
@@ -23,19 +25,80 @@ export class PaintingsService {
   constructor(
     @InjectModel(Painting)
     private paintingModel: typeof Painting,
+    @InjectModel(PaintingAttributes)
+    private paintingAttributesModel: typeof PaintingAttributes,
     private readonly storageService: StorageService
   ) {}
 
   async create(createPaintingDto: CreatePaintingDto): Promise<Painting> {
     try {
+      // Логируем данные, полученные с фронта
+      this.logger.debug(
+        'Received data from frontend:',
+        JSON.stringify(createPaintingDto, null, 2)
+      )
+
       const painting = new Painting({
         ...createPaintingDto,
-        artistId: createPaintingDto.artistId
+        artistId: createPaintingDto.artistId,
+        priority: 0
       })
-      this.logger.debug(painting, 'painting')
+      this.logger.debug(
+        'Painting data before save:',
+        JSON.stringify(painting, null, 2)
+      )
       await painting.save()
+
+      // Сохраняем связи с материалами
+      if (createPaintingDto.materials) {
+        for (const materialId of createPaintingDto.materials) {
+          await this.paintingAttributesModel.create({
+            paintingId: painting.id,
+            attributeId: materialId,
+            type: 'materialsList'
+          })
+        }
+      }
+
+      // Сохраняем связи с техниками
+      if (createPaintingDto.techniques) {
+        for (const techniqueId of createPaintingDto.techniques) {
+          await this.paintingAttributesModel.create({
+            paintingId: painting.id,
+            attributeId: techniqueId,
+            type: 'techniquesList'
+          })
+        }
+      }
+
+      // Сохраняем связи с темами
+      if (createPaintingDto.themes) {
+        for (const themeId of createPaintingDto.themes) {
+          await this.paintingAttributesModel.create({
+            paintingId: painting.id,
+            attributeId: themeId,
+            type: 'themesList'
+          })
+        }
+      }
+
+      // Получаем полные данные о картине с атрибутами
+      const fullPainting = await this.paintingModel.findOne({
+        where: { id: painting.id },
+        include: [
+          { model: Artist, attributes: ['artistName'] },
+          { model: Attributes, through: { attributes: ['type'] } }
+        ]
+      })
+
+      return fullPainting
+
       return painting
     } catch (error) {
+      this.logger.error(
+        `Error creating painting: ${error.message}`,
+        error.stack
+      )
       throw new InternalServerErrorException(
         `Error creating painting: ${error.message}`
       )
@@ -69,7 +132,6 @@ export class PaintingsService {
 
     /* filters starts */
     const parsedFilters = filters ? JSON.parse(filters) : {}
-    // this.logger.debug(parsedFilters, 'parsedFilters')
     const {
       artTypesList = [],
       colorsList = [],
@@ -149,7 +211,10 @@ export class PaintingsService {
   async findOne(id: string): Promise<Painting> {
     const options: FindOptions = {
       where: { id },
-      include: [{ model: Artist, attributes: ['artistName'] }]
+      include: [
+        { model: Artist, attributes: ['artistName'] },
+        { model: Attributes, through: { attributes: ['type'] } }
+      ]
     }
     const painting = await this.paintingModel.findOne(options)
     if (!painting) {
@@ -183,10 +248,55 @@ export class PaintingsService {
       }
     )
 
+    // Обновляем связи с материалами
+    if (painting.materials) {
+      await PaintingAttributes.destroy({
+        where: { paintingId: id, type: 'materialsList' }
+      })
+      for (const materialId of painting.materials) {
+        await PaintingAttributes.create({
+          paintingId: id,
+          attributeId: materialId,
+          type: 'materialsList'
+        })
+      }
+    }
+
+    // Обновляем связи с техниками
+    if (painting.techniques) {
+      await this.paintingAttributesModel.destroy({
+        where: { paintingId: id, type: 'techniquesList' }
+      })
+      for (const techniqueId of painting.techniques) {
+        await PaintingAttributes.create({
+          paintingId: id,
+          attributeId: techniqueId,
+          type: 'techniquesList'
+        })
+      }
+    }
+
+    // Обновляем связи с темами
+    if (painting.themes) {
+      await PaintingAttributes.destroy({
+        where: { paintingId: id, type: 'themesList' }
+      })
+      for (const themeId of painting.themes) {
+        await this.paintingAttributesModel.create({
+          paintingId: id,
+          attributeId: themeId,
+          type: 'themesList'
+        })
+      }
+    }
+
     // Теперь делаем запрос для получения обновленных данных с автором
     const updatedPainting = await this.paintingModel.findOne({
       where: { id: id },
-      include: [{ model: Artist, attributes: ['artistName'] }]
+      include: [
+        { model: Artist, attributes: ['artistName'] },
+        { model: Attributes, through: { attributes: ['type'] } }
+      ]
     })
 
     if (!updatedPainting) {
@@ -197,19 +307,31 @@ export class PaintingsService {
   }
 
   async delete(id: string): Promise<void> {
+    this.logger.debug(`Attempting to delete painting with id: ${id}`)
+
     const painting = await this.findOne(id)
+    this.logger.debug(`Found painting: ${JSON.stringify(painting, null, 2)}`)
 
     if (!painting) {
+      this.logger.error(`Painting with id ${id} not found`)
       throw new NotFoundException(`Painting with id ${id} not found`)
     }
 
     const imgUrl = painting.imgUrl
     const fileName = getFileNameFromUrl(imgUrl)
+    this.logger.debug(`Deleting file: ${fileName}`)
 
     try {
+      // Удаляем связанные записи из PaintingAttributes
+      await PaintingAttributes.destroy({
+        where: { paintingId: id }
+      })
+
       await this.storageService.deleteFile(fileName, 'paintings')
       await painting.destroy()
+      this.logger.debug(`Painting with id ${id} deleted successfully`)
     } catch (error) {
+      this.logger.error(`Error deleting painting: ${error.message}`)
       throw new InternalServerErrorException(
         `Error deleting painting: ${error.message}`
       )
@@ -219,17 +341,33 @@ export class PaintingsService {
   async deleteMany(ids: string): Promise<{ deletedPaintingCount: number }> {
     const idArray = JSON.parse(ids).map((id) => id.toString())
     let deletedPaintingCount = 0
+
     for (const id of idArray) {
       try {
         const painting = await this.findOne(id)
+        this.logger.debug(`Deleting painting with id: ${id}`)
+
+        if (!painting) {
+          this.logger.error(`Painting with id ${id} not found`)
+          continue // Пропускаем, если картина не найдена
+        }
 
         const imgUrl = painting.dataValues.imgUrl
         const fileName = getFileNameFromUrl(imgUrl)
 
+        // Удаляем связанные записи из PaintingAttributes
+        await PaintingAttributes.destroy({
+          where: { paintingId: id }
+        })
+
         await this.storageService.deleteFile(fileName, 'paintings')
         await painting.destroy()
         deletedPaintingCount++
+        this.logger.debug(`Painting with id ${id} deleted successfully`)
       } catch (error) {
+        this.logger.error(
+          `Error deleting painting with id ${id}: ${error.message}`
+        )
         throw new InternalServerErrorException(
           `Error deleting paintings: ${error.message}`
         )
